requirements.txt
html5validator==0.2.8
Markdown==3.1.1
pelican==4.8.0
Pillow==10.2.0
pre-commit
requests==2.31.0
typogrify==2.0.7
